Split earnings result formatting module
modules/earnings-results-format.ts had grown to 1446 lines and its test file
to 1578, both well past the ~800-line module guideline. Extract four cohesive
modules and mirror the split in the tests:

- earnings-results-money.ts: numeric scanning, money cues, currency/scale
  detection, parseNumber and the value formatters.
- earnings-results-metrics.ts: metric definitions, quarter-section line
  selection and label continuation matching.
- earnings-results-document.ts: html-to-text, meaningful lines, headline,
  quarter label and document currency detection.
- earnings-results-message.ts: Discord message rendering and outcome helpers.

earnings-results-format.ts keeps parseEarningsDocument and getMessageMetrics
as entry points plus the extraction engine, and re-exports the helpers other
modules already import from it, so no caller changes.

Function bodies move verbatim; overall coverage is unchanged. The one test
that mixed money primitives with identifier normalizers is split in two so
each file covers its own module.
