+++ modules/earnings-results-document.test.ts
@@ -1,0 +1,26 @@
+import {describe, expect, test} from "vitest";
+import {decodeHtmlEntities, htmlToText} from "./earnings-results-document.ts";
+
+describe("earnings result document text", () => {
+  test("removes script/style blocks with spaced closing tags", () => {
+    const text = htmlToText(`
+      <p>Revenue $10 billion</p>
+      <script>malicious()</script
+        data-ignored>
+      <style>body { color: red; }</style
+        data-ignored>
+      <p>EPS $1.00</p>
+    `);
+
+    expect(text).toContain("Revenue $10 billion");
+    expect(text).toContain("EPS $1.00");
+    expect(text).not.toContain("malicious");
+    expect(text).not.toContain("color: red");
+  });
+
+  test("decodes html entities without double-unescaping ampersands", () => {
+    expect(decodeHtmlEntities("A&amp;B &lt;tag&gt; &amp;lt;safe&amp;gt; &#36;1")).toBe(
+      "A&B <tag> &lt;safe&gt; $1",
+    );
+  });
+});
