+++ modules/earnings-results-metrics.ts
@@ -1,0 +1,288 @@
+import {getPositionedQuarterValues} from "./earnings-results-format-selection.ts";
+
+export type EarningsResultOutcome = "beat" | "inline" | "miss";
+
+export type EarningsResultMetric = {
+  currencyCode?: string | undefined;
+  estimate?: string | undefined;
+  key: string;
+  label: string;
+  numericValue?: number | undefined;
+  outcome?: EarningsResultOutcome | undefined;
+  sourceSnippet?: string | undefined;
+  value: string;
+};
+
+export type MetricValueType = "eps" | "money" | "number";
+
+export type MetricDefinition = {
+  key: string;
+  label: string;
+  patterns: RegExp[];
+  skipPattern?: RegExp;
+  valueType: MetricValueType;
+};
+
+export type MetricLineSelection = {
+  exclusive: boolean;
+  lines: string[];
+};
+
+export const earningsMetricDefinitions: MetricDefinition[] = [
+  {
+    key: "affo_per_share",
+    label: "AFFO/share",
+    patterns: [
+      /\baffo\s+(?:and\s+affo\s+)?per\s+(?:common\s+)?share\b/i,
+      /\badjusted\s+funds?\s+from\s+operations?\s+per\s+(?:common\s+)?share\b/i,
+    ],
+    valueType: "eps",
+  },
+  {
+    key: "adjusted_eps",
+    label: "Adj EPS",
+    patterns: [
+      /\badjusted\b(?:(?![.!?]\s)[^!?\n]){0,180}?(?<metricValue>-?(?:[$€£¥]\s*)?(?:\d{1,3}(?:,\d{3})+|\d+)(?:\.\d+)?)\s+per\s+(?:common\s+)?(?:diluted\s+)?share(?:\s*[-–—]\s*diluted)?\b/i,
+      /\badjusted\s+(?:\d{1,2}\s+)?(?:continuing(?:\s+operations?)?\s+)?(?:diluted\s+)?(?:earnings\s+per\s+(?:common\s+)?share|eps)\b/i,
+      /\bnon-gaap\s+(?:fully\s+)?(?:diluted\s+)?eps\b/i,
+      /\bnon-gaap\s+(?:diluted\s+)?(?:earnings\s+per\s+share|eps)\b/i,
+    ],
+    valueType: "eps",
+  },
+  {
+    key: "gaap_eps",
+    label: "EPS",
+    patterns: [
+      /\b(?:diluted\s+)?(?:earnings|net\s+income)\s+per\s+(?:common\s+)?share\b/i,
+      /\bdiluted\s+eps\b/i,
+      /\bgaap\s+(?:diluted\s+)?eps\b/i,
+      /\b(?:reported\s+)?(?:net\s+)?earnings?\b(?:(?![.!?]\s)[^!?\n]){0,180}?(?<metricValue>-?(?:[$€£¥]\s*)?(?:\d{1,3}(?:,\d{3})+|\d+)(?:\.\d+)?)\s+per\s+(?:common\s+)?(?:diluted\s+)?share(?:\s*[-–—]\s*diluted)?\b/i,
+      /(?<metricValue>\(?-?(?:[$€£¥]\s*)?(?:\d{1,3}(?:,\d{3})+|\d+)(?:\.\d+)?\)?(?:\s*(?:cents?|¢))?)\s+per\s+(?:fully\s+)?(?:common\s+)?diluted\s+share\b/i,
+      /\beps\b/i,
+    ],
+    skipPattern: /\badjusted\b|\bnon-gaap\b|\bguidance\b|\boutlook\b|\bforecast\b|\bexcept\s+(?:eps|per\s+share(?:\s+amounts?)?)\b/i,
+    valueType: "eps",
+  },
+  {
+    key: "revenue",
+    label: "Revenue",
+    patterns: [
+      /\btotal\s+revenues?(?:\s+and\s+other\s+income)?\b/i,
+      /\bnet\s+sales\b/i,
+      /\brevenues?\b/i,
+      /\bsales\b/i,
+    ],
+    skipPattern: /\bcost\s+of\b|\bdeferred\b|\bunearned\b|\bguidance\b|\boutlook\b|\bsince\s+(?:launch|inception)\b|\blife-to-date\b|\bcumulative\b|\bannuali[sz]ed\s+(?:revenue\s+)?run[-\s]*rate\b|\brevenue\s+run[-\s]*rate\b|\brevenue\s+\(expense\)|\bnon[-\s]insurance\s+warranty\s+revenue\b|\bnot\s+recognized\s+in\s+revenue\b|\bnon-cash\s+revenues?\b|\bsales\s+volumes?\b|\b(?:external\s+power|pipeline\s+gas|hydrocarbon|asset)\s+sales\b|\bproceeds\s+from\b|\bsales\s+of\s+pipeline\s+gas\b|\b(?:kbd|koebd|boepd|bpd|mboed|mmboe|bcfe|mmcf|mw|gw|kt)\b/i,
+    valueType: "money",
+  },
+  {
+    key: "net_income",
+    label: "Net income",
+    patterns: [
+      /\bnet\s+income(?!\s+per\s+(?:common\s+)?share)\b/i,
+      /\bnet\s+earnings(?!\s+per\s+(?:common\s+)?share)\b/i,
+    ],
+    // Skip income-statement subtotals and the noncontrolling-interest component so
+    // the headline "net income/earnings attributable to <company>" row wins rather
+    // than "net earnings before income tax" or "net earnings including NCI".
+    skipPattern: /\beps\b|\bbefore\s+(?:income\s+)?tax(?:es)?\b|\b(?:including|attributable\s+to)\s+noncontrolling\b/i,
+    valueType: "money",
+  },
+  {
+    key: "refinery_throughput",
+    label: "Refinery throughput",
+    patterns: [
+      /\brefinery\s+throughput\b/i,
+    ],
+    valueType: "number",
+  },
+  {
+    key: "production",
+    label: "Production",
+    patterns: [
+      /\bproduction\b/i,
+    ],
+    skipPattern: /\b(?:capacity|startup|on\s+plan|guidance|outlook|forecast)\b/i,
+    valueType: "number",
+  },
+];
+
+export function getQuarterSpecificMetricLines(lines: string[]): MetricLineSelection {
+  const startIndex = lines.findIndex(line =>
+    isQuarterSpecificSectionLine(line) || isMixedMonthQuarterResultsLine(line));
+  if (-1 === startIndex) {
+    return {
+      exclusive: false,
+      lines: [],
+    };
+  }
+
+  const exclusive = isMixedMonthQuarterResultsLine(lines[startIndex] ?? "");
+  const selectedLines: string[] = [];
+  for (let index = startIndex; index < lines.length; index++) {
+    const line = lines[index];
+    if (undefined === line) {
+      continue;
+    }
+
+    if (index > startIndex &&
+        (true === isQuarterSpecificSectionBoundary(line) ||
+        (true === exclusive && true === isMixedMonthQuarterSectionBoundary(line)))) {
+      break;
+    }
+
+    selectedLines.push(line);
+    if (selectedLines.length >= 16) {
+      break;
+    }
+  }
+
+  return {
+    exclusive,
+    lines: selectedLines,
+  };
+}
+
+function isMixedMonthQuarterResultsLine(line: string): boolean {
+  return /\bmonth\s+and\s+quarter\s+ended\b/i.test(line) ||
+    /\bquarter\s+and\s+month\s+ended\b/i.test(line);
+}
+
+function isMixedMonthQuarterSectionBoundary(line: string): boolean {
+  return /^\s*(?:the\s+)?(?:comprehensive|consolidated)\s+(?:comprehensive\s+)?(?:income\s+)?statements?\b/i.test(line) ||
+    /^\s*for\s+the\s+(?:month|year-to-date)\b/i.test(line);
+}
+
+function isQuarterSpecificSectionLine(line: string): boolean {
+  if (/\b(?:guidance|outlook|forecast)\b/i.test(line)) {
+    return false;
+  }
+
+  return /^\s*(?:for\s+)?Q[1-4]\s+(?:(?:fiscal\s+year|FY|FYE)\s*)?(?:20\d{2}|\d{2})(?:\s+(?:financial\s+overview|results?(?:\s+summary)?|earnings))?\s*:?$/i.test(line) ||
+    /^\s*(?:for\s+)?(?:the\s+)?(?:first|second|third|fourth)[\s–—-]+quarter(?:\s+(?:of\s+)?(?:(?:fiscal\s+year|FY)\s*)?(?:20\d{2}|\d{2}))?(?:\s+(?:financial\s+overview|results?(?:\s+summary)?|earnings))?\s*:?$/i.test(line);
+}
+
+function isQuarterSpecificSectionBoundary(line: string): boolean {
+  return /^\s*(?:outlook|guidance|financial\s+outlook|business\s+outlook|use\s+of\s+non-gaap|forward-looking|supplemental\s+financial\s+information)\b/i.test(line) ||
+    /^\s*(?:the\s+)?company\s+(?:raises?|updates?|reaffirms?|provides?|issues?)\b.*\b(?:guidance|outlook)\b/i.test(line) ||
+    /^\s*(?:fiscal\s+year|FY|FYE)\s*(?:20\d{2}|\d{2})\b/i.test(line) ||
+    /^\s*for\s+fiscal\s+year\s+(?:20\d{2}|\d{2})\s*:?$/i.test(line);
+}
+
+export function isPerShareOnlyNetIncomeLine(line: string): boolean {
+  const hasCombinedAggregateLabel =
+    /\bnet\s+income\b.*\band\b.*\bnet\s+income\s+per\s+(?:common\s+|diluted\s+)?share\b/i.test(line);
+  return /\bper\s+(?:common\s+|diluted\s+)?share\b/i.test(line) &&
+    false === hasCombinedAggregateLabel &&
+    false === /\b(?:trillion|billion|million|thousand)s?\b/i.test(line);
+}
+
+export function getMetricLineWithContinuation(
+  lines: string[],
+  lineIndex: number,
+  definition: MetricDefinition,
+  quarterLabel: string | undefined,
+): string {
+  const baseLine = lines[lineIndex] ?? "";
+  const positionedQuarterValues = getPositionedQuarterValues(
+    lines,
+    lineIndex,
+    quarterLabel,
+  );
+  if (0 < positionedQuarterValues.length) {
+    return [baseLine, ...positionedQuarterValues].join(" ");
+  }
+
+  const metricLines = [baseLine];
+  const isSummaryHeading = isSummaryMetricHeading(baseLine, definition);
+  for (let index = lineIndex + 1; index < lines.length && index <= lineIndex + 6; index++) {
+    const nextLine = lines[index];
+    if (undefined === nextLine) {
+      break;
+    }
+
+    if (true === isValueOnlyLine(nextLine) ||
+        true === isPerShareMetricDetailLine(baseLine, nextLine)) {
+      metricLines.push(nextLine);
+      continue;
+    }
+
+    if (false === isSummaryHeading || true === isSummaryMetricHeadingLine(nextLine)) {
+      break;
+    }
+
+    if ("money" === definition.valueType && true === isNarrativeMoneyDetailLine(nextLine)) {
+      metricLines.push(nextLine);
+      break;
+    }
+
+    if ("eps" === definition.valueType) {
+      if (true === isNarrativePerShareDetailLine(nextLine)) {
+        metricLines.push(nextLine);
+        break;
+      }
+
+      if (true === isNarrativeMoneyDetailLine(nextLine)) {
+        continue;
+      }
+    }
+
+    break;
+  }
+
+  return metricLines.join(" ");
+}
+
+function isSummaryMetricHeading(line: string, definition: MetricDefinition): boolean {
+  return line.length <= 180 &&
+    false === /[$€£¥]|\b\d+(?:[.,]\d+)?\b/.test(line) &&
+    definition.patterns.some(pattern => pattern.test(line));
+}
+
+function isSummaryMetricHeadingLine(line: string): boolean {
+  if (line.length > 180 || /[$€£¥]|\b\d+(?:[.,]\d+)?\b/.test(line)) {
+    return false;
+  }
+
+  return earningsMetricDefinitions.some(definition =>
+    definition.patterns.some(pattern => pattern.test(line))) ||
+    /\b(?:adjusted\s+ebitda|operating\s+income)\b/i.test(line);
+}
+
+function isNarrativeMoneyDetailLine(line: string): boolean {
+  return /^\s*(?:[•◦▪–—-]\s*)?(?:\(?\s*)?(?:[$€£¥]\s*)?-?\d/i.test(line) &&
+    /[$€£¥]|\b(?:trillion|billion|million|thousand)s?\b|\b(?:tn|bn|mm|[tbmk])\b/i.test(line);
+}
+
+function isNarrativePerShareDetailLine(line: string): boolean {
+  return /^\s*(?:[•◦▪–—-]\s*)?(?:\(?\s*)?(?:[$€£¥]\s*)?-?\d/i.test(line) &&
+    /\b(?:per\s+(?:common\s+)?share|eps|cents?)\b/i.test(line);
+}
+
+function isValueOnlyLine(line: string): boolean {
+  return /^[\s|$€£¥(),.\-\d%—–]+$/.test(line);
+}
+
+function isPerShareMetricDetailLine(baseLine: string, line: string): boolean {
+  return /\bper\s+(?:common\s+)?share\b/i.test(baseLine) &&
+    /^\s*(?:basic|diluted)\b/i.test(line);
+}
+
+export function hasMixedMonthQuarterColumns(lines: string[], lineIndex: number): boolean {
+  const context = lines
+    .slice(Math.max(0, lineIndex - 8), lineIndex + 1)
+    .join(" ");
+  const hasMonth = /\b(?:month|january|february|march|april|may|june|july|august|september|october|november|december)\b/i.test(context);
+  return hasMonth && /\bquarter\b/i.test(context) && /\bchange\b/i.test(context);
+}
+
+export function isNearTableNoteColumn(lines: string[], lineIndex: number): boolean {
+  for (let index = lineIndex; index >= 0 && index >= lineIndex - 5; index--) {
+    const line = lines[index];
+    if (undefined !== line && /\bnote\b/i.test(line)) {
+      return true;
+    }
+  }
+
+  return false;
+}
